Update Neo4jConfig.scala
In NEO4J, if the parameter dbms.connector.bolt. Tls_level is set to REQUIRED, when we connect NEO4J with neo4j-spark connector, the connection will report an error:Connection to the database terminated. This can happen due to network instabilities, or due to restarts of the database.So We add a parameter to control how the Config instance can be created,If the encryptionStatus variable is false, use the withoutEncryption() method to initialize the config instance,otherwise use the withEncryption() to initialize it.